08.03.2018

+++ Wpf-test/ManuallyInputWindow.xaml.cs
@@ -28,8 +28,8 @@ namespace Wpf_test
 
             var task = new TaskClass
             {
-                Mathematic = Convert.ToInt16(this.MathTextBox.Text),
-                Dispr = Convert.ToInt16(this.DisprTextBox.Text),
+                Mathematic = Convert.ToDouble(this.MathTextBox.Text),
+                Dispr = Convert.ToDouble(this.DisprTextBox.Text),
                 TimeToWork = Convert.ToInt16(this.TimeToWorkTextBox.Text),
                 Importance = Convert.ToInt16(this.ImportanceTextBox.Text),
                 IndexNumber = Convert.ToInt16(NumberTask.Text),
